% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_groupLasso_Log.R
\name{admmlasso_ls}
\alias{admmlasso_ls}
\title{Square Loss with LASSO Penalty}
\usage{
admmlasso_ls(y, X, lam, rho = 0.001)
}
\arguments{
\item{y}{The response vector with n entries}

\item{X}{An n by p design matrix}

\item{lam}{The non-negative tuning parameter value}

\item{rho}{The penalty parameter in the augmented Lagrangian}
}
\value{
This function returns a list with four elements:
\item{beta}{The vector of regression coefficient estimates, with p-1 entries}
\item{total.iterations}{Number of iterations made}
}
\description{
Compute the LASSO penalized least square solutions
}
\examples{
\dontrun{
admmlasso_ls(X=X, y=y, lam=lam, rho=1e-3)}
}

